+++ R/survival_reg_flexsurvspline.R
@@ -1,0 +1,11 @@
+#' Flexible parametric survival regression
+#'
+#' [flexsurv::flexsurvspline()] fits a flexible parametric survival model.
+#'
+#' @includeRmd man/rmd/survival_reg_flexsurvspline.md details
+#'
+#' @name details_survival_reg_flexsurvspline
+#' @keywords internal
+NULL
+
+# See inst/README-DOCS.md for a description of how these files are processed